+++ R/input.R
@@ -12,10 +12,10 @@
 #' # Text input
 #' uiinput(
 #'   tags$label("Text input"),
-#'   uitextinput("ex", type = "text", placeholder = "Enter Text")
+#'   text_input("ex", type = "text", placeholder = "Enter Text")
 #' )
 #'
-#' @seealso uitextinput
+#' @seealso text_input
 #'
 #' @export
 uiinput <- function(..., class = "") {
@@ -32,9 +32,11 @@ uiinput <- function(..., class = "") {
 #' @param type Change depending what type of input is wanted. See details for options.
 #' @param placeholder Text visible in the input when nothing is inputted.
 #' @param attribs A named list of attributes to assign to the input.
+#' @param label character with label put on the left from the input
+#' @param width The width of the input, eg. "40px"
 #'
 #' @details
-#' The following \code{type}s are allowed:
+#' The following \code{type} s are allowed:
 #' \itemize{
 #' \item{text}{The standard input}
 #' \item{textarea}{An extended space for text}
@@ -48,17 +50,22 @@ uiinput <- function(..., class = "") {
 #' \code{\link[shiny]{updateTextAreaInput}} if \code{type = "textarea"}.
 #'
 #' @examples
-#' library(shiny)
-#' library(shiny.semantic)
-#'
-#' # Text input
-#' uiinput(
-#'   tags$label("Text input"),
-#'   uitextinput("ex", type = "text", placeholder = "Enter Text")
-#' )
-#'
+#' if (interactive()) {
+#'   library(shiny)
+#'   library(shiny.semantic)
+#'   ui <- semanticPage(
+#'     uiinput(
+#'       text_input("ex", label = "Your text", type = "text", placeholder = "Enter Text")
+#'     )
+#'   )
+#'   server <- function(input, output, session) {
+#'  }
+#'  shinyApp(ui, server)
+#'  }
+#' @rdname text_input
 #' @export
-uitextinput <- function(input_id, value = "", type = "text", placeholder = NULL, attribs = list()) {
+text_input <- function(input_id, label = NULL, value = "", type = "text",
+                       placeholder = NULL, attribs = list()) {
   if (!type %in% c("text", "textarea", "password", "email", "url", "tel")) {
     stop(type, " is not a valid Semantic UI input")
   }
@@ -70,7 +77,63 @@ uitextinput <- function(input_id, value = "", type = "text", placeholder = NULL,
   }
 
   for (i in names(attribs)) input$attribs[[i]] <- attribs[[i]]
-  input
+  if (is.null(label))
+    input
+  else
+    uiinput(
+      tags$label(label),
+      input
+    )
+
+}
+
+#' Create a semantic Text Area input
+#'
+#' Create a text area input control for entry of unstructured text values.
+#'
+#' @param inputId Input name. Reactive value is available under \code{input[[input_id]]}.
+#' @param label character with label put above the input
+#' @param value Pass value if you want to have default text.
+#' @param width The width of the input, eg. "40px"
+#' @param placeholder Text visible in the input when nothing is inputted.
+#'
+#' @examples
+#' if (interactive()) {
+#' ui <- semanticPage(
+#'   textAreaInput("a", "Area:", width = "200px"),
+#'   verbatimTextOutput("value")
+#' )
+#' server <- function(input, output, session) {
+#'   output$value <- renderText({ input$a })
+#' }
+#' shinyApp(ui, server)
+#' }
+#' @export
+textAreaInput <- function(inputId, label, value = "", width = NULL, placeholder = NULL) {
+  shiny::div(
+    class = "ui form",
+    style = if (!is.null(width)) glue::glue("width: {shiny::validateCssUnit(width)};"),
+    shiny::div(class = "field",
+               if (!is.null(label)) tags$label(label, `for` = inputId),
+               text_input(inputId, value,
+                          placeholder = placeholder, type = "textarea")
+    )
+  )
+}
+
+#' @param inputId Input name. The same as \code{input_id}.
+#' @rdname text_input
+#' @export
+textInput <- function(inputId, label, value = "", width = NULL,
+                      placeholder = NULL, type = "text") {
+  shiny::div(
+    class = "ui form",
+    style = if (!is.null(width)) glue::glue("width: {shiny::validateCssUnit(width)};"),
+    shiny::div(class = "field",
+               if (!is.null(label)) tags$label(label, `for` = inputId),
+               text_input(inputId, value, placeholder = placeholder, type = type)
+    )
+  )
 }
 
 #' Create Semantic UI Numeric Input
@@ -78,7 +141,7 @@ uitextinput <- function(input_id, value = "", type = "text", placeholder = NULL,
 #' This creates a default numeric input using Semantic UI. The input is available
 #' under \code{input[[input_id]]}.
 #'
-#' @param input_id Input input_id Reactive value is available under \code{input[[input_id]]}.
+#' @param input_id Input name. Reactive value is available under \code{input[[input_id]]}.
 #' @param value Initial value of the numeric input.
 #' @param min Minimum allowed value.
 #' @param max Maximum allowed value.
@@ -88,22 +151,19 @@ uitextinput <- function(input_id, value = "", type = "text", placeholder = NULL,
 #' @param icon Icon or label attached to numeric input.
 #' @param placeholder Inner input label displayed when no value is specified.
 #' @param ... Unused.
+#' @param label character with label
 #'
 #' @details
 #' The inputs are updateable by using \code{\link{updateNumericInput}}.
-#'
+#' @rdname numeric_input
 #' @examples
 #' library(shiny)
 #' library(shiny.semantic)
 #'
-#' # Text input
-#' uiinput(
-#'   tags$label("Numeric Input"),
-#'   uinumericinput("ex", 10)
-#' )
+#' numeric_input("ex", "Select number", 10)
 #'
 #' @export
-uinumericinput <- function(input_id, value, min = NA, max = NA, step = NA,
+numeric_input <- function(input_id, label, value, min = NA, max = NA, step = NA,
                            type = NULL, icon = NULL, placeholder = NULL, ...) {
   if (!is.numeric(value) & !grepl("^\\d*(\\.\\d*|)$", value)) stop("Non-numeric input detected")
 
@@ -114,10 +174,13 @@ uinumericinput <- function(input_id, value, min = NA, max = NA, step = NA,
   if (!is.null(icon)) {
     type <- paste(type, "icon")
   }
-  shiny::div(
-    class = paste("ui", type, "input"),
-    input_tag,
-    icon
+  shiny::div(class = "field",
+    if (!is.null(label)) tags$label(label, `for` = input_id),
+    shiny::div(
+      class = paste("ui", type, "input"),
+      input_tag,
+      icon
+    )
   )
 }
 
@@ -130,16 +193,15 @@ uinumericinput <- function(input_id, value, min = NA, max = NA, step = NA,
 #' @param max Maximum allowed value.
 #' @param step Interval to use when stepping between min and max.
 #' @param width The width of the input.
-#' @param ... Other parameters passed to \code{\link{uinumericinput}} like \code{type} or \code{icon}.
+#' @param ... Other parameters passed to \code{\link{numeric_input}} like \code{type} or \code{icon}.
+#' @rdname numeric_input
 #' @export
-numericInput <- function(inputId, label, value, min = NA, max = NA, step = NA, width = NULL, ...) {
+numericInput <- function(inputId, label, value,
+                         min = NA, max = NA, step = NA, width = NULL, ...) {
   shiny::div(
     class = "ui form",
     style = if (!is.null(width)) glue::glue("width: {shiny::validateCssUnit(width)};"),
-    shiny::div(class = "field",
-               if (!is.null(label)) tags$label(label, `for` = inputId),
-               uinumericinput(inputId, value, min, max, step, ...)
-    )
+    numeric_input(inputId, label, value, min, max, step, ...)
   )
 }
 
@@ -160,6 +222,7 @@ update_numeric_input <- function(session, input_id, label = NULL, value = NULL,
                             min = min, max = max, step = step)
 }
 
+#' @param inputId the same as \code{input_id}
 #' @export
 #' @rdname update_numeric_input
 updateNumericInput <- shiny::updateNumericInput