+++ R/rating.R
@@ -1,13 +1,4 @@
-#' Extract icon name
-#'
-#' @param icon uiicon object
-#'
-#' @return character with icon name
-extract_icon_name <- function(icon) {
-  gsub(" icon", "", icon$attribs$class)
-}
-
-#' Rating
+#' Rating Input.
 #'
 #' Crates rating component
 #'
@@ -15,15 +6,29 @@ extract_icon_name <- function(icon) {
 #' @param label the contents of the item to display
 #' @param value initial rating value
 #' @param max maximum value
-#' @param icon character with name of the icon or \code{\link{uiicon}()} that is
+#' @param icon character with name of the icon or \code{\link{icon}()} that is
 #' an element of the rating
 #' @param color character with colour name
 #' @param size character with legal semantic size, eg. "medium", "huge", "tiny"
 #'
 #' @return rating object
-#'
+#' @examples
+#' if (interactive()) {
+#'   library(shiny)
+#'   library(shiny.semantic)
+#'   ui <- shinyUI(
+#'     semanticPage(
+#'       rating_input("rate", "How do you like it?", max = 5,
+#'                    icon = "heart", color = "yellow"),
+#'     )
+#'   )
+#'   server <- function(input, output) {
+#'     observeEvent(input$rate,{print(input$rate)})
+#'   }
+#'   shinyApp(ui = ui, server = server)
+#' }
 #' @export
-rating <- function(input_id, label = "", value = 0, max = 3, icon = "star",
+rating_input <- function(input_id, label = "", value = 0, max = 3, icon = "star",
                    color = "yellow", size = "") {
   if (!(size %in% SIZE_LEVELS)) {
     warning("Size value not supported.")
@@ -45,7 +50,8 @@ rating <- function(input_id, label = "", value = 0, max = 3, icon = "star",
 
 #' Update rating
 #'
-#' check \code{rating} to learn more.
+#' Change the value of a rating input on the client. Check
+#' \code{rating_input} to learn more.
 #'
 #' @param session shiny object with session info
 #'
@@ -55,14 +61,8 @@ rating <- function(input_id, label = "", value = 0, max = 3, icon = "star",
 #'
 #' @export
 #' @rdname update_rating
-update_rating <- function(session, input_id, label = NULL, value = NULL) {
+update_rating_input <- function(session, input_id, label = NULL, value = NULL) {
   message <- list(label = label, value = value)
   message <- message[!vapply(message, is.null, FUN.VALUE = logical(1))]
   session$sendInputMessage(input_id, message)
-}
-
-#' @export
-#' @rdname update_rating
-updateRating <- function(session, inputId, label = NULL, value = NULL) {
-  update_rating(session, inputId, label, value)
 }