+++ R/slider.R
@@ -2,7 +2,7 @@
 #'
 #' @description
 #' This creates a slider input using Semantic UI. Slider is already initialized and
-#' available under \code{input[[input_id]]}.
+#' available under \code{input[[input_id]]}. Use Range for range of values.
 #'
 #' @param input_id Input name. Reactive value is available under \code{input[[input_id]]}.
 #' @param value The initial value to be selected for the sldier (lower value if using range).
@@ -14,7 +14,7 @@
 #' @details
 #' Use \code{\link{update_slider}} to update the slider/range within the shiny session.
 #'
-#' @rdname uislider
+#' @rdname slider
 #'
 #' @examples
 #' if (interactive()) {
@@ -27,7 +27,7 @@
 #'     semanticPage(
 #'       title = "Slider example",
 #'       tags$br(),
-#'       uislider("slider", 10, 0, 20),
+#'       slider_input("slider", 10, 0, 20),
 #'       p("Selected value:"),
 #'       textOutput("slider")
 #'     )
@@ -44,7 +44,7 @@
 #'      semanticPage(
 #'        title = "Range example",
 #'        tags$br(),
-#'        uirange("range", 10, 15, 0, 20),
+#'        range_input("range", 10, 15, 0, 20),
 #'        p("Selected values:"),
 #'        textOutput("range")
 #'     )
@@ -62,33 +62,60 @@
 #' \url{https://fomantic-ui.com/modules/slider.html} for preset classes.
 #'
 #' @export
-uislider <- function(input_id, value, min, max, step = 1, class = NULL) {
+slider_input <- function(input_id, value, min, max, step = 1, class = NULL) {
   div(
     id = input_id, class = paste("ui slider", class),
     `data-min` = min, `data-max` = max, `data-step` = step, `data-start` = value
   )
 }
 
-#' @rdname uislider
+#' @param inputId Input name.
+#' @param label Display label for the control, or NULL for no label.
+#' @param width character with width of slider.
+#' @param ... additional arguments
+#' @rdname slider
+#' @export
+sliderInput <- function(inputId, label, min, max, value, step = 1, width = NULL, ...) {
+  check_extra_arguments(list(...))
+  form(
+    style = if (!is.null(width)) glue::glue("width: {shiny::validateCssUnit(width)};"),
+    label(label),
+    slider_input(inputId, value, min, max, step = step)
+  )
+}
+
+#' @rdname slider
 #' @param value2 The initial upper value of the slider.
 #'
 #' @export
-uirange <- function(input_id, value, value2, min, max, step = 1, class = NULL) {
+range_input <- function(input_id, value, value2, min, max, step = 1, class = NULL) {
   div(
     id = input_id, class = paste("ui range slider", class),
     `data-min` = min, `data-max` = max, `data-step` = step, `data-start` = value, `data-end` = value2
   )
 }
 
+#' @param inputId Input name.
+#' @rdname slider
+#' @export
+rangeInput <- function(inputId, label, min, max, value, step = 1, width = NULL, ...) {
+  check_extra_arguments(list(...))
+  form(
+    style = if (!is.null(width)) glue::glue("width: {shiny::validateCssUnit(width)};"),
+    label(label),
+    range_input(inputId, value, min, max, step = step)
+  )
+}
+
 #' Update slider Semantic UI component
 #'
-#' Change the value of a \code{\link{uislider}} input on the client.
+#' Change the value of a \code{\link{slider_input}} input on the client.
 #'
 #' @param session The \code{session} object passed to function given to \code{shinyServer}.
 #' @param input_id The id of the input object
 #' @param value The value to be selected for the sldier (lower value if using range).
 #'
-#' @seealso uislider
+#' @seealso slider_input
 #'
 #' @rdname update_slider
 #' @export
@@ -98,10 +125,26 @@ update_slider <- function(session, input_id, value) {
 }
 
 #' @rdname update_slider
-#' @param value2 The upper value of the slider.
+#' @param value2 The upper value of the range.
 #'
 #' @export
 update_range <- function(session, input_id, value, value2) {
   message <- list(value = jsonlite::toJSON(c(value, value2)))
   session$sendInputMessage(input_id, message)
+}
+
+#' @param inputId Input name.
+#' @param ... additional arguments
+#' @rdname update_slider
+#' @export
+updateSliderInput <- function(session, inputId, value,  ...) {
+  check_extra_arguments(list(...))
+  update_slider(session, inputId, value)
+}
+
+#' @param inputId Input name.
+#' @rdname update_slider
+#' @export
+updateRangeInput <- function(session, inputId, value, value2) {
+  update_range(session, inputId, value, value2)
 }